+++ src/main/java/vehicles/ElectricCar.java
@@ -6,7 +6,7 @@ import components.Tyres;
 
 public class ElectricCar extends Vehicle{
 
-    public ElectricCar(Engine engine, GearBox gearbox, Tyres tyres, int price, String colour, String make, String model, Boolean electric, Boolean hybrid) {
+    public ElectricCar(Engine engine, GearBox gearbox, Tyres tyres, double price, String colour, String make, String model, Boolean electric, Boolean hybrid) {
         super(engine, gearbox, tyres, price, colour, make, model, electric, hybrid);
     }
 }
